1. 添加运行环境相关的代码在基础类和启动器中 2. 升级打包插件，符合新的运维需求 4. 日志现在支持预备不同环境的配置文件 5. 一些工具类的 Bugfix

+++ mybatis-support/src/main/java/com/joindata/inf/common/support/mybatis/support/properties/DataSourceProperties.java
@@ -46,10 +46,10 @@ public class DataSourceProperties extends BaseDataSourceProperties
     protected long maxWait;
 
     /** 检测关闭空闲连接的间隔时间毫秒数 */
-    protected int timeBetweenEvictionRunsMillis;
+    protected long timeBetweenEvictionRunsMillis;
 
     /** 最小生存时间毫秒数 */
-    protected int minEvictableIdleTimeMillis;
+    protected long minEvictableIdleTimeMillis;
 
     /** 验证语句 */
     protected String validationQuery;
@@ -141,23 +141,23 @@ public class DataSourceProperties extends BaseDataSourceProperties
     }
 
     @DisconfFileItem(name = "jdbc.timeBetweenEvictionRunsMillis", associateField = "timeBetweenEvictionRunsMillis")
-    public int getTimeBetweenEvictionRunsMillis()
+    public long getTimeBetweenEvictionRunsMillis()
     {
         return timeBetweenEvictionRunsMillis;
     }
 
-    public void setTimeBetweenEvictionRunsMillis(int timeBetweenEvictionRunsMillis)
+    public void setTimeBetweenEvictionRunsMillis(long timeBetweenEvictionRunsMillis)
     {
         this.timeBetweenEvictionRunsMillis = timeBetweenEvictionRunsMillis;
     }
 
     @DisconfFileItem(name = "jdbc.minEvictableIdleTimeMillis", associateField = "minEvictableIdleTimeMillis")
-    public int getMinEvictableIdleTimeMillis()
+    public long getMinEvictableIdleTimeMillis()
     {
         return minEvictableIdleTimeMillis;
     }
 
-    public void setMinEvictableIdleTimeMillis(int minEvictableIdleTimeMillis)
+    public void setMinEvictableIdleTimeMillis(long minEvictableIdleTimeMillis)
     {
         this.minEvictableIdleTimeMillis = minEvictableIdleTimeMillis;
     }
